version: Test for VerQueryValueA (try 2)
This patch supersedes both '[2/2] version: Test for VerQueryValueA'
and '[1/2] version: Fix wrong length reported when selecting a block.
(resend 2)', as it is a better tests, and proves the fix to not work
fully.

The patch shows a difference in handling between wine's VerQueryValue,
and windows' VerQueryValue when selecting a 'BLOCK' object.

Implemented pointer checking and commented the binary structure as
suggested by nsivov from #winehackers.