wined3d: Do not rotate WINED3D_SWAP_EFFECT_DISCARD swapchains.
RenderDoc/Qt creates swapchains with 2 buffers, a front buffer and a back
buffer, and the DISCARD swap effect. It doesn't redraw the back buffer after a
Present(), seemingly expecting that the back buffer will only be copied to the
front buffer, and otherwise remain unmodified. Incidentally, CEF seems to
behave in a similar way.