windowscodecs/tiff: Skip frames marked with SUBFILETYPE(0x1).
This is important for RAW formats that use TIFF container, but mark
directories as 'reduced-size'. Those files should be processed with
appropriate RAW decoder, instead of displaying auxiliary frame with
the TIFF decoder.

Unfortunately libtiff skips old tag OSUBFILETYPE == 0xff,
it's not available via TIFFGetField() and SUBFILETYPE is not synthesized
from it. On Windows both tags have the same effect on frames
being skipped by the decoder.

Signed-off-by: Nikolay Sivov <nsivov@codeweavers.com>