light.msstyles: Make toolbar button background transparent at the center.
Qt5 applications draw a white close glyph on top of the toolbar button background for its Qt5QWindowIcon
window class. Before this patch, the toolbar button background was also white, so the close glyph was
not distinguishable from the background. On Windows 10, the toolbar button background has a light blue
color with transparency at the center, so let's make the center of the toolbar button background of
light.msstyles transparent as well.